requirements.txt
appnope==0.1.0
apptools==4.4.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0rc1
bleach==1.5.0
certifi==2017.4.17
configobj==5.0.6
configparser==3.5.0
cx-Freeze==5.0.2
cycler==0.10.0
Cython==0.25.2
decorator==4.0.11
entrypoints==0.2.3
enum34==1.1.6
envisage==4.6.0
funcsigs==1.0.2
functools32==3.2.3.post2
h5py==2.7.0
html5lib==0.9999999
ipykernel==4.6.1
ipython==5.4.1
ipython-genutils==0.2.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
Keras==2.0.5
keras-vis==0.3.1
Markdown==2.2.0
MarkupSafe==1.0
matplotlib==2.0.2
mayavi==4.5.0
mistune==0.7.4
mock==2.0.0
mpi4py==2.0.0
nbconvert==5.2.1
nbformat==4.3.0
networkx==1.11
nltk==3.2.4
notebook==5.0.0
numexpr==2.6.2
numpy==1.13.0
olefile==0.44
pandas==0.20.2
pandocfilters==1.4.1
pathlib2==2.3.0
pbr==3.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
ProDy==1.8
prompt-toolkit==1.0.14
protobuf==3.3.0
ptyprocess==0.5.2
pyface==5.1.0
Pygments==2.2.0
PyInstaller==3.2.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
scandir==1.5
scikit-image==0.13.0
scikit-learn==0.18.1
scipy==0.19.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
subprocess32==3.2.7
tables==3.4.2
tensorflow==1.2.0
terminado==0.6
testpath==0.3.1
Theano==0.9.0
tornado==4.5.1
tqdm==4.14.0
traitlets==4.3.2
traits==4.6.0
traitsui==5.1.0
wcwidth==0.1.7
Werkzeug==0.12.2
wxPython==3.0.2.0
wxPython-common==3.0.2.0
